move bench_offline_throughput to sglang.benchmark; repoint importers
